LPS-95285 Refactor FileUtil out of internal for child projects
