+++ modules/sdk/project-templates/project-templates/src/main/java/com/liferay/project/templates/WorkspaceUtil.java
@@ -14,7 +14,6 @@
 
 package com.liferay.project.templates;
 
-import com.liferay.project.templates.internal.util.FileUtil;
 import com.liferay.project.templates.internal.util.StringUtil;
 
 import java.io.File;
